Split external account links out of components/common.js
