+++ src/components/common.js
@@ -1,6 +1,5 @@
 import * as clipboard from 'clipboard-polyfill/text'
 import _ from 'lodash/fp'
-import * as qs from 'qs'
 import { Fragment, useState } from 'react'
 import FocusLock from 'react-focus-lock'
 import { b, div, h, img, label, span } from 'react-hyperscript-helpers'
@@ -8,21 +7,15 @@ import { ButtonPrimary } from 'src/components/common/buttons'
 import Clickable from 'src/components/common/Clickable'
 import { IdContainer } from 'src/components/common/IdContainer'
 import Link from 'src/components/common/Link'
-import { spinnerOverlay } from 'src/components/common/spinners'
 import { icon } from 'src/components/icons'
 import { TextInput } from 'src/components/input'
 import Modal from 'src/components/Modal'
 import { MiniSortable } from 'src/components/table'
 import scienceBackground from 'src/images/science-background.jpg'
-import { Ajax } from 'src/libs/ajax'
 import { getEnabledBrand, isRadX } from 'src/libs/brand-utils'
 import colors from 'src/libs/colors'
 import { getConfig } from 'src/libs/config'
 import { withErrorReporting } from 'src/libs/error'
-import * as Nav from 'src/libs/nav'
-import { notify } from 'src/libs/notifications'
-import { useOnMount } from 'src/libs/react-utils'
-import { authStore } from 'src/libs/state'
 import * as Utils from 'src/libs/utils'
 
 
@@ -62,67 +55,6 @@ export const methodLink = config => {
   return sourceRepo === 'agora' ?
     `${getConfig().firecloudUrlRoot}/?return=${getEnabledBrand().queryName}#methods/${methodNamespace}/${methodName}/${methodVersion}` :
     `${getConfig().dockstoreUrlRoot}/workflows/${methodPath}:${methodVersion}`
-}
-
-export const ShibbolethLink = ({ button = false, children, ...props }) => {
-  const nihRedirectUrl = `${window.location.origin}/${Nav.getLink('profile')}?nih-username-token=<token>`
-  return h(button ? ButtonPrimary : Link, _.merge({
-    href: `${getConfig().shibbolethUrlRoot}/login?${qs.stringify({ 'return-url': nihRedirectUrl })}`,
-    ...(button ? {} : { style: { display: 'inline-flex', alignItems: 'center' } }),
-    ...Utils.newTabLinkProps
-  }, props), [children, icon('pop-out', { size: 12, style: { marginLeft: '0.2rem' } })])
-}
-
-export const FrameworkServiceLink = ({ linkText, provider, redirectUrl, button = false, ...props }) => {
-  const [href, setHref] = useState()
-
-  useOnMount(() => {
-    const loadAuthUrl = withErrorReporting('Error getting Fence Link', async () => {
-      const result = await Ajax().User.getFenceAuthUrl(provider, redirectUrl)
-      setHref(result.url)
-    })
-    loadAuthUrl()
-  })
-
-  return !!href ?
-    h(button ? ButtonPrimary : Link, {
-      href,
-      ...(button ? {} : { style: { display: 'inline-flex', alignItems: 'center' } }),
-      ...Utils.newTabLinkProps,
-      ...props
-    }, [linkText, icon('pop-out', { size: 12, style: { marginLeft: '0.2rem' } })]) :
-    h(Fragment, [linkText])
-}
-
-export const UnlinkFenceAccount = ({ linkText, provider }) => {
-  const [isModalOpen, setIsModalOpen] = useState(false)
-  const [isUnlinking, setIsUnlinking] = useState(false)
-
-  return div({ style: { display: 'inline-flex' } }, [
-    h(Link, { onClick: () => { setIsModalOpen(true) } }, [linkText]),
-    isModalOpen && h(Modal, {
-      title: 'Confirm unlink account',
-      onDismiss: () => setIsModalOpen(false),
-      okButton: h(ButtonPrimary, {
-        onClick: _.flow(
-          withErrorReporting('Error unlinking account'),
-          Utils.withBusyState(setIsUnlinking)
-        )(async () => {
-          await Ajax().User.unlinkFenceAccount(provider.key)
-          authStore.update(_.set(['fenceStatus', provider.key], {}))
-          setIsModalOpen(false)
-          notify('success', 'Successfully unlinked account', {
-            message: `Successfully unlinked your account from ${provider.name}`,
-            timeout: 30000
-          })
-        })
-      }, 'OK')
-    }, [
-      div([`Are you sure you want to unlink from ${provider.name}?`]),
-      div({ style: { marginTop: '1rem' } }, ['You will lose access to any underlying datasets. You can always re-link your account later.']),
-      isUnlinking && spinnerOverlay
-    ])
-  ])
 }
 
 export const FocusTrapper = ({ children, onBreakout, ...props }) => {
