+++ src/libs/link-expiration-alerts.js
@@ -2,7 +2,7 @@ import { addDays, differenceInDays, parseJSON } from 'date-fns/fp'
 import _ from 'lodash/fp'
 import { Fragment, useEffect, useState } from 'react'
 import { h } from 'react-hyperscript-helpers'
-import { FrameworkServiceLink, ShibbolethLink, UnlinkFenceAccount } from 'src/components/common'
+import { FrameworkServiceLink, ShibbolethLink, UnlinkFenceAccount } from 'src/components/external-account-links'
 import { getEnabledBrand } from 'src/libs/brand-utils'
 import * as Nav from 'src/libs/nav'
 import allProviders from 'src/libs/providers'
